pyproject.toml
[project]
name = "crewai_skills_crew"
version = "0.1.0"
description = "One agent, any task — skills discovered and loaded at runtime via crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=1.10.1",
    "python-dotenv>=1.0.0"
]

[project.scripts]
crewai_skills_crew = "crewai_skills_crew.main:run"
run_crew = "crewai_skills_crew.main:run"
train = "crewai_skills_crew.main:train"
replay = "crewai_skills_crew.main:replay"
test = "crewai_skills_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
